Update file names and default exports
