+++ packages/jss-plugin-vendor-prefixer/.size-snapshot.json
@@ -1,13 +1,13 @@
 {
   "dist/jss-plugin-vendor-prefixer.js": {
-    "bundled": 17480,
-    "minified": 5616,
-    "gzipped": 2183
+    "bundled": 17314,
+    "minified": 5593,
+    "gzipped": 2167
   },
   "dist/jss-plugin-vendor-prefixer.min.js": {
-    "bundled": 17480,
-    "minified": 5616,
-    "gzipped": 2183
+    "bundled": 17314,
+    "minified": 5593,
+    "gzipped": 2167
   },
   "dist/jss-plugin-vendor-prefixer.cjs.js": {
     "bundled": 1375,
