+++ packages/react-jss/src/JssContext.js
@@ -2,7 +2,9 @@
 import React, {type Context} from 'react'
 import type {Context as JssContextValue} from './types'
 
-export const JssContext: Context<JssContextValue> = React.createContext({
+const JssContext: Context<JssContextValue> = React.createContext({
   sheetOptions: {classNamePrefix: ''},
   disableStylesGeneration: false
 })
+
+export default JssContext;
