+++ packages/react-jss/src/index.js
@@ -1,12 +1,12 @@
 // @flow
-import {withStyles} from './with-styles'
-import {createUseStyles} from './create-use-styles'
+import withStyles from './withStyles'
+import createUseStyles from './createUseStyles'
 
 export {ThemeProvider, withTheme, createTheming} from 'theming'
-export {JssProvider} from './JssProvider'
-export {jss} from './jss'
+export {default as JssProvider} from './JssProvider'
+export {default as jss} from './jss'
 export {SheetsRegistry, createGenerateId} from 'jss'
-export {JssContext} from './JssContext'
+export {default as JssContext} from './JssContext'
 
 export {withStyles, createUseStyles}
