+++ packages/react-jss/src/jss.js
@@ -2,4 +2,4 @@
 import {create} from 'jss'
 import preset from 'jss-preset-default'
 
-export const jss = create(preset())
+export default create(preset())
