+++ packages/react-jss/src/utils/index-counter.js
@@ -12,7 +12,7 @@
  */
 let index = 0
 
-function getIndex() {
+const getIndex = () => {
   return index++
 }
