+++ packages/react-jss/src/utils/managers.js
@@ -5,7 +5,7 @@ import type {Context} from '../types'
 
 const defaultManagers = new Map()
 
-function getManager(context: Context, managerId: number) {
+const getManager = (context: Context, managerId: number) => {
   // If `managers` map is present in the context, we use it in order to
   // let JssProvider reset them when new response has to render server-side.
   if (context.managers) {
