+++ packages/react-jss/src/utils/mergeClasses.js
